uat: extract framework to separate module
