+++ purebred.cabal
@@ -144,6 +144,7 @@ test-suite uat
   hs-source-dirs:      test
   ghc-options:         -Wall -threaded
   main-is:             TestUserAcceptance.hs
+  other-modules:       UAT
   default-language:    Haskell2010
   build-depends:       base
                      , purebred-email >= 0.1 && < 0.2
